requirements.txt
apache-airflow==1.10.10
cryptography==2.9
requests==2.23.0
urllib3==1.25.9
beautifulsoup4==4.9.0
pandas==0.25.3
